Fix UI changes
