+++ fireapp/test/presentation/shift_request/shift_request_view_model_test.dart
@@ -10,10 +10,7 @@ import 'package:rxdart/rxdart.dart';
 
 import 'shift_request_view_model_test.mocks.dart';
 
-@GenerateNiceMocks([
-  MockSpec<ShiftRequestRepository>(),
-])
-
+@GenerateMocks([ShiftRequestRepository])
 void main() {
   group('ShiftRequestViewModel', () {
     late MockShiftRequestRepository shiftRequestRepository;
@@ -31,20 +28,23 @@ void main() {
           assetClass: "Class1",
           startTime: DateTime.now(),
           endTime: DateTime.now().add(Duration(hours: 1)),
-          shiftVolunteers: [], // Add mock shift volunteers if needed
+          shiftVolunteers: [],
         ),
-        // Add more mock shift requests if needed
       ];
 
       when(shiftRequestRepository.getShiftRequestsByRequestID(any))
           .thenAnswer((_) async => shiftRequests);
 
-      viewModel.loadShiftRequests("mockRequestId");
+      final emittedStates = [];
+      final subscription = viewModel.shiftRequests.listen(emittedStates.add);
+
+      viewModel.loadShiftRequests(requestId: "mockRequestId");
       await Future.delayed(Duration(milliseconds: 100));
 
-      final shiftRequestsState = viewModel.shiftRequests as BehaviorSubject<RequestState<List<ShiftRequest>>>;
-      expect(shiftRequestsState.value, isA<SuccessRequestState<List<ShiftRequest>>>());
-      expect((shiftRequestsState.value as SuccessRequestState).result, isNotEmpty);
+      expect(emittedStates.last, isA<SuccessRequestState<List<ShiftRequest>>>());
+      expect((emittedStates.last as SuccessRequestState<List<ShiftRequest>>).result, isNotEmpty);
+
+      subscription.cancel();
     });
 
     test('deletes shift assignment successfully', () async {
@@ -62,20 +62,19 @@ void main() {
       subscription.cancel();
     });
 
-  test('updates shift by position successfully', () async {
-    when(shiftRequestRepository.updateShiftByPosition(any, any, any))
-        .thenAnswer((_) async => null);
+    test('updates shift by position successfully', () async {
+      when(shiftRequestRepository.updateShiftByPosition(any, any, any))
+          .thenAnswer((_) async => null);
 
-    final emittedStates = [];
-    final subscription = viewModel.updateState.listen(emittedStates.add);
+      final emittedStates = [];
+      final subscription = viewModel.updateState.listen(emittedStates.add);
 
-    viewModel.updateShiftByPosition(1, 1, 1);
-    await Future.delayed(Duration(milliseconds: 100));
+      viewModel.updateShiftByPosition(1, 1, 1);
+      await Future.delayed(Duration(milliseconds: 100));
 
-    expect(emittedStates.last, isA<SuccessRequestState<void>>());
+      expect(emittedStates.last, isA<SuccessRequestState<void>>());
 
-    subscription.cancel();
+      subscription.cancel();
     });
-
   });
 }
